+++ src/main/java/software/aoc/day03/a/JoltageStrategy.java
@@ -2,5 +2,5 @@ package software.aoc.day03.a;
 
 @FunctionalInterface
 public interface JoltageStrategy {
-    int calculate(Battery battery);
+    long calculate(Battery battery);
 }
